solutions/p0993/solution_test.go
package solution

import (
	"testing"
)

func Test_isCousins(t *testing.T) {
	type args struct {
		root *TreeNode
		x    int
		y    int
	}
	tests := []struct {
		name string
		args args
		want bool
	}{
		// TODO: Add test cases
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := isCousins(tt.args.root, tt.args.x, tt.args.y); got != tt.want {
				t.Errorf("isCousins() = %v, want %v", got, tt.want)
			}
		})
	}
}

func Test_helper(t *testing.T) {
	type args struct {
		root   *TreeNode
		depth  int
		parent int
		m      map[int][2]int
	}
	tests := []struct {
		name string
		args args
	}{
		// TODO: Add test cases.
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			helper(tt.args.root, tt.args.depth, tt.args.parent, tt.args.m)
		})
	}
}
